# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server e-commerce server.",
        "title": "Backend GO",
        "contact": {
            "name": "João Cabral",
            "url": "https://github.com/C4BRALL",
            "email": "cabral047dev@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:9874",
    "paths": {
        "/seller": {
            "post": {
                "description": "Create a new seller",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seller"
                ],
                "parameters": [
                    {
                        "description": "seller request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSellerInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "seller created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
        "/seller/signin": {
            "post": {
                "description": "Generate token jwt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seller"
                ],
                "parameters": [
                    {
                        "description": "seller request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetJwtInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetJwtOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
        "/seller/{email}": {
            "get": {
                "description": "Get a seller by email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seller"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "E-mail of the seller",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Seller"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
        "/sellers/all": {
            "get": {
                "description": "Get all sellers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seller"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "page_number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Seller"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateSellerInput": {
            "type": "object",
            "properties": {
                "document": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "dto.GetJwtInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.GetJwtOutput": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
        "entity.Seller": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "document": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/enums.Status"
                },
                "stores": {
                    "description": "Não é necessário gorm:\"foreignKey:SellerID\"",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Store"
                    }
                },
                "type": {
                    "$ref": "#/definitions/enums.UserType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entity.Store": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sellerID": {
                    "description": "Adicione uma chave estrangeira aqui",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/enums.Status"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "enums.Status": {
            "type": "integer",
            "enum": [
                1,
                0
            ],
            "x-enum-varnames": [
                "Active",
                "Inactive"
            ]
        },
        "enums.UserType": {
            "type": "string",
            "enum": [
                "customers",
                "seller"
            ],
            "x-enum-varnames": [
                "Customer",
                "Salesperson"
            ]
        },
        "handlers.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}